Folder structure rework + fixed deleting all boards and adding a new one crashes the app
