+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginActivity">
+    tools:context=".authentication.LoginActivity">
 
     <TextView
         android:id="@+id/project_name"
